experiments/asr_backbone_comparison/pyproject.toml
[project]
name = "asr-backbone-comparison"
version = "0.1.0"
description = "CTC-based ASR encoder backbone comparison on Ukrainian Common Voice"
requires-python = ">=3.10"
dependencies = [
    # PyTorch — install with CUDA wheel separately (see README)
    "torch==2.4.1",
    "torchaudio==2.4.1",
    # Mamba SSM (requires CUDA build tools; install after torch)
    "mamba-ssm",
    "causal-conv1d",
    # Data
    "datasets>=2.20",
    "soundfile>=0.12",
    "librosa>=0.10",
    # Metrics
    "jiwer>=3.0",
    # Config + utilities
    "pyyaml>=6.0",
    "numpy>=1.24",
    "matplotlib>=3.8",
    # RWKV-block is installed separately via setup_rwkv.sh
]

[project.scripts]
run-experiment = "scripts.run_experiment:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asr_exp"]

[tool.uv]
# CUDA index for torch — adjust cu124 to match your CUDA version
index-url = "https://download.pytorch.org/whl/cu124"

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
